servicedesk-agent/pyproject.toml
[project]
name = "servicedesk-agent"
version = "0.1.0"
description = "Multi-Agent Service Desk system using LangGraph and Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.14",
    "langgraph>=0.2.60",
    "langchain-google-genai>=2.0.8",
    "langchain-core>=0.3.28",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.6",
    "mypy>=1.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]
